+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## next
+
+- HTTP requests: Log response body on error.
+
 ## 0.5.0
 
 2020-12-24